requirements.txt
pyyaml
numpy==1.24.4
pandas
botocore==1.35.4
PyAthena==3.9.0
SQLAlchemy==2.0.35
scikit-learn
skforecast
statsmodels
pmdarima
optuna
matplotlib
lightgbm==4.5.0
fastapi
uvicorn
catboost==1.2.7
xgboost==2.1.3
mlforecast==0.15.0
scipy==1.10.1
awswrangler
python-dotenv
boto3
glob2